introdução a implementação da lógica de calculo necessário na camada service

+++ desafio-zg/src/main/java/com/desafio_zg/domain/model/UserTrade.java
@@ -29,7 +29,7 @@ public class UserTrade {
     private String instrument;
     private String especificacao;
 
-    private Double quantidade;
+    private int quantidade;
     private Double preco;
     private Double valorTotal;
